adds test for crud endpoints and swagger module

+++ src/snaps/snaps.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
 import { CreateSnapDto } from './dto/create-snap.dto';
 import { Snap } from './entities/snap.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { UUID } from 'typeorm/driver/mongodb/bson.typings';
+import { isUUID } from 'class-validator';
 
 @Injectable()
 export class SnapsService {
@@ -12,20 +13,31 @@ export class SnapsService {
   ) {}
 
   async create(createSnapDto: CreateSnapDto) {
-    const snap = this.snapRepository.create(createSnapDto);
-    return this.snapRepository.save(snap);
+    const snap = await this.snapRepository.create(createSnapDto);
+    return { data: await this.snapRepository.save(snap) };
   }
 
-  findAll() {
-    return this.snapRepository.find();
+  async findAll() {
+    return { data: await this.snapRepository.find() };
   }
 
-  findOne(id: UUID) {
-    return this.snapRepository.findOne({ where: { id: id } });
+  async findOne(id: UUID) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid UUID');
+    }
+    const snap = await this.snapRepository.findOne({ where: { id: id } })
+    if (!snap) {
+      throw new NotFoundException('Snap not found');
+    }
+    return { data: snap };
   }
 
   async remove(id: UUID) {
-    const snap = await this.findOne(id);
-    return this.snapRepository.remove(snap);
+    const snap = await this.snapRepository.findOne({ where: { id: id } });
+    if (!snap) {
+      throw new NotFoundException('Snap not found');
+    }
+    await this.snapRepository.remove(snap);
+    return "Snap deleted successfully"
   }
 }
